+++ FrontEnd/src/app/components/gestion-client/gestion-client.component.ts
@@ -68,9 +68,26 @@ export class GestionClientComponent implements OnInit {
   }
 
   onDelete(id: string) {
-    this.clientService.deleteClient(id).subscribe((response) => {
-      console.log(response);
-      this.getAllClient();
+    Swal.fire({
+      title: 'Are you sure?',
+      text: 'You will not be able to recover this client!',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#d33',
+      cancelButtonColor: '#3085d6',
+      confirmButtonText: 'Yes, delete it!',
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.clientService.deleteClient(id).subscribe((response) => {
+          console.log(response);
+          this.getAllClient();
+          Swal.fire(
+            'Deleted!',
+            'Client has been deleted.',
+            'success'
+          );
+        });
+      }
     });
   }
